+++ src/App.js
@@ -9,9 +9,9 @@ import Mypage from "./page/Mypage.js";
 import Home from "./page/Home.js";
 import Main from "./page/Main.js";
 import Detail from "./page/Detail.js";
-import Socialkakao from "./Socialkakao.js";
-import Socialnaver from "./Socialnaver.js";
-import Socialgoogle from "./Socialgoogle.js";
+import Socialkakao from "./page/tmp/Socialkakao.js";
+import Socialnaver from "./page/tmp/Socialnaver.js";
+import Socialgoogle from "./page/tmp/Socialgoogle.js";
 import Search from "./page/Search.js";
 import Movie from "./page/Movie.js";
 import Tv from "./page/Tv.js";
